Move convertToTimeString to helpers and use it
