+++ libs/utils/src/lib/helpers.ts
@@ -222,3 +222,24 @@ export function capitalize(text: string) {
 export function unique(array: string[]) {
   return Array.from(new Set(array));
 }
+
+/**
+ * Convert a number to time string
+ * @param time In milliseconds
+ */
+export function convertToTimeString(time: number) {
+  let hour: number, minute: number, second: number;
+  second = Math.floor(time / 1000);
+  minute = Math.floor(second / 60);
+  second = second % 60;
+  hour = Math.floor(minute / 60);
+  minute = minute % 60;
+  const day = Math.floor(hour / 24);
+  hour = hour % 24;
+  hour += day * 24;
+  const dayStr = day > 0 ? `${day}d` : '';
+  const hourStr = hour > 0 ? `${hour}h` : '';
+  const minuteStr = minute > 0 ? `${minute}min` : '';
+  const secondStr = second > 0 ? `${second}s` : '';
+  return `${dayStr}${hourStr}${minuteStr}${secondStr}`;
+}
