+++ libs/utils/src/lib/pipes/duration.pipe.ts
@@ -1,5 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import { NgModule } from '@angular/core';
+import { convertToTimeString } from '../helpers';
 
 @Pipe({
   name: 'duration'
@@ -23,29 +24,7 @@ export class DurationPipe implements PipeTransform {
     ) {
       return placeholder;
     }
-
-    return this.convertToTimeString(value);
-  }
-
-  convertToTimeString(time: number) {
-
-    let hour: number, minute: number, second: number;
-
-    second = Math.floor(time / 1000);
-    minute = Math.floor(second / 60);
-    second = second % 60;
-    hour = Math.floor(minute / 60);
-    minute = minute % 60;
-    const day = Math.floor(hour / 24);
-    hour = hour % 24;
-    hour += day * 24;
-
-    const dayStr = day > 0 ? `${day}d` : '';
-    const hourStr = hour > 0 ? `${hour}h` : '';
-    const minuteStr = minute > 0 ? `${minute}min` : '';
-    const secondStr = second > 0 ? `${second}s` : '';
-
-    return `${dayStr}${hourStr}${minuteStr}${secondStr}`;
+    return convertToTimeString(value);
   }
 }
